Final commit
this version of the code first fetches all the videos of the playlis, then after processing them it writes the data into the csv file.
this approach can be used for smaller playlists with less than 100 videos, bigger playlists might have a risk of taking too much time which can lead to loss of data if the script is inturrupted in between.